Create Try1-v1
My stages are not matching up at all. The Canal always changes by >0.1m and that throws off the Marsh because all of that extra is getting dumped into it. They both start out in the neighborhood but then they slowly diverge, both of them on the high side, interestingly.